requirements.txt
# =============================================================================
# Threat Modeling AI - Dependências (dataset real + treino YOLO + notebooks)
# Instalar no venv e usar kernel "Python (threat-modeling-ai)" nos notebooks.
# =============================================================================

# Core científico e imagens
numpy>=1.24.0,<3.0.0
pandas>=2.0.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# YOLO (Ultralytics) e PyTorch
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Grafo (STRIDE / pipeline)
networkx>=3.1

# Dataset e config
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0

# Roboflow (download dataset em formato YOLO)
roboflow>=1.1.0

# Validação e tipos
pydantic>=2.0.0

# Jupyter (notebooks + kernel)
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Testes
pytest>=7.4.0
pytest-cov>=4.1.0

# Linter e formatter (Ruff)
ruff>=0.2.0

# Opcionais (para depois: UI, diagramas sintéticos, diagramas reais + LLM)
streamlit>=1.28.0
diagrams>=0.23.0
openai>=1.3.0
pytm>=1.3.0
graphviz>=0.20.0

# LLM & RAG Integration
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-text-splitters
chromadb>=0.4.0
pypdf
tiktoken
